+++ virtusize/src/test/java/com/virtusize/android/parsers/I18nLocalizationJsonParserTest.kt
@@ -8,7 +8,6 @@ import com.google.common.truth.Truth.assertThat
 import com.virtusize.android.TestUtils
 import com.virtusize.android.data.local.VirtusizeLanguage
 import com.virtusize.android.data.remote.I18nLocalization
-import com.virtusize.android.fixtures.TestFixtures
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
@@ -31,35 +30,6 @@ class I18nLocalizationJsonParserTest {
             getExpectedI18nLocalization(
                 virtusizeLanguage = VirtusizeLanguage.EN,
                 localizedContext = context,
-            )
-
-        assertThat(actualI18nLocalization).isEqualTo(expectedI18nLocalization)
-    }
-
-    @Test
-    fun parseI18N_emptyJsonData_shouldReturnExpectedObject() {
-        val actualI18nLocalization =
-            I18nLocalizationJsonParser(context, VirtusizeLanguage.EN).parse(
-                TestFixtures.EMPTY_JSON_DATA,
-            )
-
-        val expectedI18nLocalization =
-            I18nLocalization(
-                VirtusizeLanguage.EN,
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
-                "",
             )
 
         assertThat(actualI18nLocalization).isEqualTo(expectedI18nLocalization)
@@ -125,6 +95,7 @@ class I18nLocalizationJsonParserTest {
             localizedContext.getString(com.virtusize.android.core.R.string.inpage_one_size_will_fit_result_text),
             localizedContext.getString(com.virtusize.android.core.R.string.inpage_multi_size_comparison_text),
             localizedContext.getString(com.virtusize.android.core.R.string.inpage_will_fit_result_text),
+            localizedContext.getString(com.virtusize.android.core.R.string.inpage_will_not_fit_result_text),
             localizedContext.getString(com.virtusize.android.core.R.string.inpage_will_not_fit_result_text),
             localizedContext.getString(com.virtusize.android.core.R.string.inpage_body_data_empty_text),
         )