+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Changes
+- Fix: issue with no size found text
+- Fix: Virtusize button state when loading a new product
+
 ### 2.12.13
 - Fix: External Link redirect issue